Clean up addition/subtraction code
